src/main/java/bf/hsawadogo/customer-service/src/main/java/bf/hsawadogo/customerservice/mappers/CustomerMapper.java
package bf.hsawadogo.customerservice.mappers;

import bf.hsawadogo.customerservice.dtos.CustomerDTO;
import bf.hsawadogo.customerservice.entities.Customer;
import org.modelmapper.ModelMapper;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.stream.Collectors;

@Service
public class CustomerMapper {
    private ModelMapper modelMapper = new ModelMapper();

    public CustomerDTO toDto(Customer customer) {
        return modelMapper.map(customer, CustomerDTO.class);
    }

    public List<CustomerDTO> toDto(List<Customer> customers) {
        return customers.stream().map(customer -> modelMapper.map(customer, CustomerDTO.class)).collect(Collectors.toList());
    }

    public Customer toEntity(CustomerDTO customerDTO) {
        return modelMapper.map(customerDTO, Customer.class);
    }

    public List<Customer> toEntity(List<CustomerDTO> customerDTOs) {
        return customerDTOs.stream().map(customerDTO -> modelMapper.map(customerDTO, Customer.class)).collect(Collectors.toList());
    }
}
